+++ pycm/pycm_compare.py
@@ -3,7 +3,7 @@
 from __future__ import division
 from .pycm_error import pycmCompareError
 from .pycm_output import *
-from .pycm_util import *
+from .utils import *
 from .pycm_param import *
 from .pycm_obj import ConfusionMatrix
 import os