+++ pycm/pycm_handler.py
@@ -4,7 +4,7 @@ from __future__ import division
 from .pycm_class_func import class_statistics
 from .pycm_error import pycmVectorError, pycmMatrixError
 from .pycm_overall_func import overall_statistics
-from .pycm_util import *
+from .utils import *
 from .pycm_param import *
 import json
 import types