+++ pycm/pycm_obj.py
@@ -9,7 +9,7 @@ from .pycm_class_func import F_calc, IBA_calc, TI_calc, NB_calc, sensitivity_ind
 from .pycm_overall_func import weighted_kappa_calc, weighted_alpha_calc, alpha2_calc, brier_score_calc, log_loss_calc
 from .pycm_distance import DistanceType, DISTANCE_MAPPER
 from .pycm_output import *
-from .pycm_util import *
+from .utils import *
 from .pycm_param import *
 from .pycm_ci import __CI_overall_handler__, __CI_class_handler__
 import os