+++ pycm/pycm_output.py
@@ -3,7 +3,7 @@
 from __future__ import division
 from functools import partial
 from .pycm_param import *
-from .pycm_util import rounder, sort_char_num
+from .utils import rounder, sort_char_num
 import webbrowser