+++ pycm/pycm_overall_func.py
@@ -6,7 +6,7 @@ import operator as op
 from functools import reduce
 from .pycm_interpret import *
 from .pycm_ci import kappa_SE_calc, CI_calc, SE_calc
-from .pycm_util import complement
+from .utils import complement
 
 
 def log_loss_calc(