+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Test system modified
 - `Python 3.12` added to `test.yml`
 - Warning and error messages updated
+- `pycm_util.py` renamed to `utils.py`
 ## [4.0] - 2023-06-07
 ### Added
 - `pycmMultiLabelError` class